添加dear imgui支持

+++ RecastDemo/Include/NavMeshTesterTool.h
@@ -84,7 +84,7 @@ class NavMeshTesterTool : public SampleTool
 	bool m_eposSet;
 	bool m_perfTest;
 	bool m_perfTestFindPoly;
-	float m_perfTestCount;
+	int m_perfTestCount;
 	float m_perfTestResult;
 
 	int m_pathIterNum;
